requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
Django==4.2.8
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
motor==3.3.2
orjson==3.9.10
passlib==1.7.4
prompt-toolkit==3.0.43
protobuf==4.21.12
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
PyMySQL==1.1.0
pyodbc==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.14
sqlparse==0.4.4
starlette==0.32.0.post1
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
uvicorn==0.25.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
